[CEN-874] Add Award Period e2e test
