+++ test/tests/bpdHbAwardPeriodV2.js
@@ -13,16 +13,3 @@ export function GetAwardPeriodSuccess(baseUrl, params) {
     console.log(`Attempted ${res.headers['Ocp-Apim-Operationid']}. Unsuccessful. Response status ${res.status}. Please check trace ${res.headers['Ocp-Apim-Trace-Location']}`)
   }
 }
-
-export function GetAwardPeriodSuccessV1(baseUrl, params) {
-
-  const res = http.get(
-    `${baseUrl}/bpd/hb/award-periods`,
-    params
-  );
-
-  const isSuccessful = check(res, { 'Success': (r) => r.status === 200 });
-  if (!isSuccessful) {
-    console.log(`Attempted ${res.headers['Ocp-Apim-Operationid']}. Unsuccessful. Response status ${res.status}. Please check trace ${res.headers['Ocp-Apim-Trace-Location']}`)
-  }
-}
